+++ lib/rails_health_checks/configuration.rb
@@ -4,10 +4,11 @@ module RailsHealthChecks
   class ConfigurationError < StandardError; end
 
   class Configuration
-    BUILT_IN_NAMES = %i[database cache sidekiq solid_queue good_job resque disk memory http].freeze
+    BUILT_IN_NAMES = %i[database cache redis sidekiq solid_queue good_job resque disk memory http].freeze
 
     attr_writer :checks
     attr_accessor :timeout, :cache_duration, :allowed_ips, :token,
+                  :redis_url,
                   :sidekiq_queue_size, :solid_queue_job_count, :good_job_latency,
                   :resque_queue_size, :disk_warn_threshold, :disk_critical_threshold, :disk_path,
                   :memory_threshold, :http_url, :http_expected_status, :http_headers
@@ -20,6 +21,7 @@ module RailsHealthChecks
       @allowed_ips = nil
       @token = nil
       @authenticate_block = nil
+      @redis_url = nil
       @sidekiq_queue_size = nil
       @solid_queue_job_count = nil
       @good_job_latency = nil