+++ spec/generators/rails_health_checks/initializer_generator_spec.rb
@@ -70,6 +70,10 @@ RSpec.describe RailsHealthChecks::Generators::InitializerGenerator do
     expect(initializer_content).to include("config.disk_critical_threshold")
   end
 
+  it "includes redis check option" do
+    expect(initializer_content).to include("config.redis_url")
+  end
+
   it "includes memory check option" do
     expect(initializer_content).to include("config.memory_threshold")
   end